Restructures global state
